Add files via upload
A bit more about the project

This project is covering a basic Api automation test framework and some sample test cases. Although, we can surely enhance it more with lots of robust features.

Positive Cases : The implement automated tests are the following -

getBookings: Testing at least 2 existing bookings are returned in the response.
getBooking: Testing the data returned for an existing booking matches.
createBooking: Testing the bookings can be created.
Negative Cases : The implement automated tests are the following -

A room cannot be booked more than once for a given date.
The check-out date must be greater than the check-in date.